broke al the things that make config out into their own classes etc..sigh
