+++ src/Ocelot/Configuration/Creator/IAggregatesCreator.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface IAggregatesCreator
+    {
+        List<ReRoute> Aggregates(FileConfiguration fileConfiguration, List<ReRoute> reRoutes);
+    }
+}
