+++ src/Ocelot/Configuration/Creator/IConfigurationCreator.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface IConfigurationCreator
+    {
+        InternalConfiguration InternalConfiguration(FileConfiguration fileConfiguration, List<ReRoute> reRoutes);
+    }
+}
