+++ src/Ocelot/Configuration/Creator/IDynamicsCreator.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface IDynamicsCreator
+    {
+        List<ReRoute> Dynamics(FileConfiguration fileConfiguration);
+    }
+}
