+++ src/Ocelot/Configuration/Creator/ILoadBalancerOptionsCreator.cs
@@ -1,0 +1,9 @@
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface ILoadBalancerOptionsCreator
+    {
+        LoadBalancerOptions CreateLoadBalancerOptions(FileLoadBalancerOptions options);
+    }
+}
