+++ src/Ocelot/Configuration/Creator/IReRoutesCreator.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface IReRoutesCreator
+    {
+        List<ReRoute> ReRoutes(FileConfiguration fileConfiguration);
+    }
+}
