+++ src/Ocelot/Configuration/Creator/LoadBalancerOptionsCreator.cs
@@ -1,0 +1,17 @@
+using Ocelot.Configuration.File;
+
+namespace Ocelot.Configuration.Creator
+{
+
+    public class LoadBalancerOptionsCreator : ILoadBalancerOptionsCreator
+    {
+        public LoadBalancerOptions CreateLoadBalancerOptions(FileLoadBalancerOptions options)
+        {
+            return new LoadBalancerOptionsBuilder()
+                .WithType(options.Type)
+                .WithKey(options.Key)
+                .WithExpiryInMs(options.Expiry)
+                .Build();
+        }
+    }
+}
